changed timer return values to double to decrease delay between screen refreshes

+++ timer.h
@@ -10,7 +10,7 @@ class Timer
 		std::chrono::time_point<std::chrono::high_resolution_clock> cur;
 	public:
 		Timer();
-		int getTime();
+		double getTime();
 		void update();
 };
